fix supporter icon in profile missing its background
also it doesn't show up on profiles without supporter anymore
