+++ osu.Game/Users/UserPanel.cs
@@ -16,8 +16,8 @@ using osu.Game.Overlays;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Graphics.UserInterface;
 using osu.Framework.Graphics.Cursor;
-using osu.Game.Graphics.Backgrounds;
 using osu.Game.Graphics.Containers;
+using osu.Game.Overlays.Profile;
 
 namespace osu.Game.Users
 {
@@ -220,53 +220,5 @@ namespace osu.Game.Users
         {
             new OsuMenuItem("View Profile", MenuItemType.Highlighted, ViewProfile),
         };
-
-        private class SupporterIcon : CircularContainer
-        {
-            private readonly Box background;
-
-            public SupporterIcon()
-            {
-                Masking = true;
-                Children = new Drawable[]
-                {
-                    new Box { RelativeSizeAxes = Axes.Both },
-                    new CircularContainer
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        RelativeSizeAxes = Axes.Both,
-                        Scale = new Vector2(0.8f),
-                        Masking = true,
-                        Children = new Drawable[]
-                        {
-                            background = new Box { RelativeSizeAxes = Axes.Both },
-                            new Triangles
-                            {
-                                TriangleScale = 0.2f,
-                                ColourLight = OsuColour.FromHex(@"ff7db7"),
-                                ColourDark = OsuColour.FromHex(@"de5b95"),
-                                RelativeSizeAxes = Axes.Both,
-                                Velocity = 0.3f,
-                            },
-                        }
-                    },
-                    new SpriteIcon
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        RelativeSizeAxes = Axes.Both,
-                        Icon = FontAwesome.fa_heart,
-                        Scale = new Vector2(0.45f),
-                    }
-                };
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                background.Colour = colours.Pink;
-            }
-        }
     }
 }
